app/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-allauth==0.63.6
django-cors-headers==4.4.0
dj-rest-auth==6.0.0
environs==11.0.0
gunicorn==22.0.0
psycopg[binary,pool]==3.2.1
requests==2.32.3